Database

Description

Database module header

Documentation

initialiseDB :: IO Connection

This is a function is for initialising the database and to create the table if the tables do not exists yet

insertDB :: Connection -> [HolidayRecord] -> IO ()

This function will insert the holiday records into the database

insertLB :: Connection -> [HolidayRecord] -> IO ()

This function will insert the country records into the dsatabase

insertSB :: Connection -> [HolidayRecord] -> IO ()

This function will insert the country_holidays records into the dsatabase

queryDB :: Connection -> String -> IO [[SqlValue]]

This function will select all the holidays of a given country

selectHolidaysInDateRange :: Connection -> String -> String -> IO [String]

This function will select all the holidays in the date specified of a given country

getNames :: Connection -> IO [String]

This function will call all the names on the database.

getLocalNames :: Connection -> Bool -> IO [String]

This function will call all the names on the database.

recordToSqlValues :: HolidayRecord -> [SqlValue]

This function is used to convert from Haskell datatypes to SQL datatypes toSql function used from HDBC to do the convert for all values to SQL in the database

holidayToSqlValues :: HolidayRecord -> [SqlValue]

This function is used to convert from Haskell datatypes to SQL datatypes toSql function used from HDBC to do the convert for all values to SQL in the database

prepareInsertRecordStmt :: Connection -> IO Statement

This function performs an action (ie. to insert the values into holidays)

prepareSelectRecordStma :: Connection -> IO Statement

This function performs an action (ie. to fetch the values from holidays)

saveHolidayRecord :: [HolidayRecord] -> Connection -> IO ()

Thus function for saving all the records into the database passing the executeMany function in HDBC library then execute a list of lists of values and it preforms an action which dose not return anything it just preforms that sql statement

savecountriesRecord :: [HolidayRecord] -> Connection -> IO ()

This function saves the country record

sqlRowToString :: [[SqlValue]] -> [String]

This function converts into Row into the string

getUnprocessedSQLHolidays :: Connection -> IO [HolidayRecord]

This function is to retrieve all the SQLs on the database

convertToJSON :: Connection -> IO String

This function converts SQL into JSON